use uint8_t for tileID

+++ Source/drlg_l1.cpp
@@ -1955,7 +1955,7 @@ void FillChambers()
 void FloodTransparancyValues()
 {
 	int yy = 16;
-	int tileID = 13;
+	uint8_t tileID = 13;
 	for (int j = 0; j < DMAXY; j++) {
 		int xx = 16;
 		for (int i = 0; i < DMAXX; i++) {
