+++ Source/drlg_l4.cpp
@@ -1193,7 +1193,7 @@ void LoadDiabQuads(bool preflag)
 void FloodTransparancyValues()
 {
 	int yy = 16;
-	int tileID = 6;
+	uint8_t tileID = 6;
 	for (int j = 0; j < DMAXY; j++) {
 		int xx = 16;
 		for (int i = 0; i < DMAXX; i++) {
